# perf(sim): Replace allocating player‑prediction loop with zero‑GC implementation