--- a/build/projaimbot.lua
+++ b/build/projaimbot.lua
@@ -454,6 +454,8 @@ local function CreateMove(uCmd)
 
 	local weaponInfo = GetProjectileInformation(pWeapon:GetPropInt("m_iItemDefinitionIndex"))
 	local vecHeadPos = pLocal:GetAbsOrigin() + pLocal:GetPropVector("localdata", "m_vecViewOffset[0]")
+	-- real charge for bows / stickies / Beggar etc.
+	local charge_time = GetCharge(pWeapon)
 
 	local best_target = GetClosestEntityToFov(pLocal, vecHeadPos, players, bAimAtTeamMates)
 	if not best_target.index then
@@ -479,7 +481,7 @@ local function CreateMove(uCmd)
 		return nil
 	end
 
-	local velocity_vector = weaponInfo:GetVelocity(0)
+	local velocity_vector = weaponInfo:GetVelocity(charge_time) -- use real charge
 	local forward_speed = math.sqrt(velocity_vector.x ^ 2 + velocity_vector.y ^ 2)
 
 	local detonate_time = pWeapon:GetWeaponID() == E_WeaponBaseID.TF_WEAPON_PIPEBOMBLAUNCHER and 0.7 or 0
@@ -550,7 +552,6 @@ local function CreateMove(uCmd)
 		return
 	end
 
-	local charge_time = GetCharge(pWeapon)
 	local gravity = client.GetConVar("sv_gravity") * weaponInfo:GetGravity(charge_time)
 	local angle = math_utils.SolveBallisticArc(vecHeadPos, predicted_target_pos, forward_speed, gravity)
 	if not angle then
@@ -595,7 +596,7 @@ local function CreateMove(uCmd)
 	elseif bIsSandvich then
 		uCmd.buttons = uCmd.buttons | IN_ATTACK2
 		bAttack = true -- special case for sandvich
-	else -- generic weapons
+	else         -- generic weapons
 		if wep_utils.CanShoot() then
 			if settings.autoshoot then
 				uCmd.buttons = uCmd.buttons | IN_ATTACK