--- a/src/main.lua
+++ b/src/main.lua
@@ -411,6 +411,8 @@ local function CreateMove(uCmd)
 
 	local weaponInfo = GetProjectileInformation(pWeapon:GetPropInt("m_iItemDefinitionIndex"))
 	local vecHeadPos = pLocal:GetAbsOrigin() + pLocal:GetPropVector("localdata", "m_vecViewOffset[0]")
+	-- real charge for bows / stickies / Beggar etc.
+	local charge_time = GetCharge(pWeapon)
 
 	local best_target = GetClosestEntityToFov(pLocal, vecHeadPos, players, bAimAtTeamMates)
 	if not best_target.index then
@@ -436,7 +438,7 @@ local function CreateMove(uCmd)
 		return nil
 	end
 
-	local velocity_vector = weaponInfo:GetVelocity(0)
+	local velocity_vector = weaponInfo:GetVelocity(charge_time) -- use real charge
 	local forward_speed = math.sqrt(velocity_vector.x ^ 2 + velocity_vector.y ^ 2)
 
 	local detonate_time = pWeapon:GetWeaponID() == E_WeaponBaseID.TF_WEAPON_PIPEBOMBLAUNCHER and 0.7 or 0
@@ -467,7 +469,7 @@ local function CreateMove(uCmd)
 
 	local vecMins, vecMaxs = weaponInfo.m_vecMins, weaponInfo.m_vecMaxs
 	local trace = engine.TraceHull(vecWeaponFirePos, predicted_target_pos, vecMins, vecMaxs, MASK_SHOT_HULL, shouldHit)
-	local is_visible = trace and trace.fraction >= 0.9
+	local is_visible = trace and (trace.fraction >= 0.9 or trace.entity == pTarget)
 
 	local bIsHuntsman = pWeapon:GetWeaponID() == E_WeaponBaseID.TF_WEAPON_COMPOUND_BOW
 
@@ -503,11 +505,10 @@ local function CreateMove(uCmd)
 
 	-- Recheck trace for final prediction
 	trace = engine.TraceHull(vecWeaponFirePos, predicted_target_pos, vecMins, vecMaxs, MASK_SHOT_HULL, shouldHit)
-	if not trace or trace.fraction < 0.9 then
+	if not trace or (trace.fraction < 0.9 and trace.entity ~= pTarget) then
 		return
 	end
 
-	local charge_time = GetCharge(pWeapon)
 	local gravity = client.GetConVar("sv_gravity") * weaponInfo:GetGravity(charge_time)
 	local angle = math_utils.SolveBallisticArc(vecHeadPos, predicted_target_pos, forward_speed, gravity)
 	if not angle then
@@ -552,7 +553,7 @@ local function CreateMove(uCmd)
 	elseif bIsSandvich then
 		uCmd.buttons = uCmd.buttons | IN_ATTACK2
 		bAttack = true -- special case for sandvich
-	else -- generic weapons
+	else         -- generic weapons
 		if wep_utils.CanShoot() then
 			if settings.autoshoot then
 				uCmd.buttons = uCmd.buttons | IN_ATTACK