--- a/src/utils/weapon_utils.lua
+++ b/src/utils/weapon_utils.lua
@@ -3,41 +3,33 @@ local wep_utils = {}
 ---@type table<integer, integer>
 local ItemDefinitions = {}
 
-local old_weapon, lastFire, nextAttack = 0, 0, 0
-
+local old_weapon, lastFire, nextAttack
 local function GetLastFireTime(weapon)
 	return weapon:GetPropFloat("LocalActiveTFWeaponData", "m_flLastFireTime")
 end
-
 local function GetNextPrimaryAttack(weapon)
 	return weapon:GetPropFloat("LocalActiveWeaponData", "m_flNextPrimaryAttack")
 end
-
 --- https://www.unknowncheats.me/forum/team-fortress-2-a/273821-canshoot-function.html
 function wep_utils.CanShoot()
 	local player = entities:GetLocalPlayer()
 	if not player then
 		return false
 	end
-
 	local weapon = player:GetPropEntity("m_hActiveWeapon")
 	if not weapon or not weapon:IsValid() then
 		return false
 	end
-
 	if weapon:GetPropInt("LocalWeaponData", "m_iClip1") == 0 then
 		return false
 	end
-
-	--[[local lastfiretime = GetLastFireTime(weapon)
-
-	if lastFire ~= lastfiretime or weapon:GetIndex() ~= old_weapon then
+	local lastfiretime = GetLastFireTime(weapon)
+	if lastFire ~= lastfiretime or weapon ~= old_weapon then
 		lastFire = lastfiretime
 		nextAttack = GetNextPrimaryAttack(weapon)
 	end
-
-	old_weapon = weapon:GetIndex()]]
-	return GetNextPrimaryAttack(weapon) <= globals.CurTime()
+	old_weapon = weapon
+	return nextAttack <= globals.CurTime()
 end
 
 do