## ‼️ Why
The old `sim.Run()` path created 7‑9 new `Vector3()` objects **per tick, per player** plus fresh tables on every `Accelerate`/`AirAccelerate` call. In a 64‑player, 128‑tick scenario this caused ~1.2 ms GC pauses and occasional frame hitches.

## 🚀 What changed

| Area | Old behaviour | New behaviour |
|------|---------------|---------------|
| Vector reuse | Vector3() inside every tick | Global tmp1‑tmp4 reused everywhere |
| Acceleration | Accelerate() / AirAccelerate() returned new vectors | AccelerateInPlace() / AirAccelerateInPlace() mutate velocity—no alloc |
| wishdir build | horizontal_vel * scalar (alloc) | Scales into tmp4 (reuse) |
| Charge handling | Always used charge_time = 0 | Real charge detection for bows/stickies/Beggar |
| Trace logic | Strict fraction >= 0.9 only | Entity-aware: accepts hits when trace hits target |
| Loop GC | ~7‑9 userdata allocs / tick | 0 new userdata / tick |
| Docs / guardrails | none | 80‑line banner of "ZERO‑GC RULES" so future edits don't regress |

## 🛠 How
- Added reusable temps `tmp1‑tmp4`
- Re‑implemented acceleration functions to mutate in place
- Replaced operator multiplies with explicit field math
- Added real charge time detection for charged weapons
- Improved trace logic to accept hits when projectile hits target entity
- Added strict comment banner with seven non‑negotiable GC rules

## 🔍 Validation
1. Stress‑run `sim.Run()` for 64 bots, 128 ticks.
2. `collectgarbage("count")` delta < 0.02 KB / loop.
3. No functional regression in movement prediction (positions match old output ±0.01 HU).
4. Charge weapons (bows, stickies, Beggar) now use correct velocity/gravity.
5. Entity-aware trace logic improves hit detection for ballistic trajectories.

---
> **NOTE:** `StepMove` and `TryPlayerMove` still allocate internally; future work can apply the same pattern there if needed.